+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v8.30.1 (unreleased)
 ### Implementation changes and bug fixes
 - [PR #1877](https://github.com/rqlite/rqlite/pull/1875): Move to no-checkpoint-on-database-close.
+- [PR #1880](https://github.com/rqlite/rqlite/pull/1880): Improve DB mode checks.
 
 ## v8.30.0 (September 5th 2024)
 ### New features