+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v8.30.1 (unreleased)
+## v8.30.1 (September 9th 2024)
 ### Implementation changes and bug fixes
 - [PR #1877](https://github.com/rqlite/rqlite/pull/1877): Move to no-checkpoint-on-database-close.
 - [PR #1880](https://github.com/rqlite/rqlite/pull/1880): Improve DB mode checks.